[MOV] sale_stock: split sale_order.py
