+++ addons/sale_stock/models/__init__.py
@@ -2,9 +2,10 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import account_move
-from . import res_company
-from . import sale_order
-from . import res_config_settings
-from . import stock
 from . import product_template
+from . import res_company
+from . import res_config_settings
 from . import res_users
+from . import sale_order
+from . import sale_order_line
+from . import stock
